Push data to GitHub

Push all currently tracked data files to GitHub. Only files identical to those already on GitHub (by md5sum hash) will not be transferred. Otherwise, assumes local version should overwrite existing GitHub version. Create a new release if you do not want to overwrite previous GitHub versions when pushing.

pb_push(repo = guess_repo(), tag = "latest", overwrite = TRUE,
  manifest = ".manifest.json")

Arguments

repo
Name of the repo on GitHub (owner/repo, i.e. cboettig/piggyback). By default will guess the current repository's GitHub origin.

tag
name of release/tag on GitHub to which data assets will be attached. Default is to use the latest available release.

overwrite
should existing files be overwritten when hashes do not match? default TRUE.

manifest
name of the local manifest file. Note: A leading dot (i.e. indicating a hidden file) in the manifest name will be removed from the name used on the GitHub asset list.

Details

Will only upload tracked files, as identified by the local manifest. Add files to tracking with pb_track first.

Examples

# NOT RUN {
pb_push()
# }