Upload data to an existing release

NOTE: you must first create a release if one does not already exists.

pb_upload(
  file,
  repo = guess_repo(),
  tag = "latest",
  name = NULL,
  overwrite = "use_timestamps",
  use_timestamps = NULL,
  show_progress = TRUE,
  .token = get_token(),
  dir = "."
)

Arguments

file

path to file to be uploaded

repo

Repository name in format "owner/repo". Will guess the current repo if not specified.

tag

tag for the GitHub release to which this data should be attached.

name

name for uploaded file. If not provided will use the basename of file (i.e. filename without directory)

overwrite

overwrite any existing file with the same name already attached to the on release? Default behavior is based on timestamps, only overwriting those files which are older.

use_timestamps

DEPRECATED.

show_progress

logical, show a progress bar be shown for uploading? Defaults to TRUE.

.token

GitHub authentication token, see [gh::gh_token()]

dir

directory relative to which file names should be based.

Examples

if (FALSE) {
# Needs your real token to run

readr::write_tsv(mtcars,"mtcars.tsv.xz")
pb_upload("mtcars.tsv.xz", "cboettig/piggyback-tests")
}