Download data from an existing release

Download data from an existing release

pb_download(file = NULL, dest = ".", repo = guess_repo(),
  tag = "latest", overwrite = TRUE, ignore = "manifest.json")

Arguments

file
name or vector of names of files to be downloaded. If NULL, all assets attached to the release will be downloaded.

dest
name of vector of names of where file should be downloaded. Should be a directory or a list of filenames the same length as file vector. Can include paths to files, but any directories in that path must already exist.

repo
Repository name in format "owner/repo". Will guess the current repo if not specified.

tag
tag for the GitHub release to which this data is attached

overwrite
default TRUE, should any local files of the same name be overwritten?

ignore
a list of files to ignore (if downloading "all" because file=NULL). by default, "manifest.json" is ignored as this file is created and used only for avoiding redundant file transfer by pb_push and pb_pull