List releases in repository

This function retrieves information about all releases attached to a given repository.

Usage

pb_releases(
  repo = guess_repo(),
  .token = gh::gh_token(),
  verbose = getOption("piggyback.verbose", default = TRUE)
)

Arguments

repo

GitHub repository specification in the form of "owner/repo", if not specified will try to guess repo based on current working directory.

.token

a GitHub API token, defaults to gh::gh_token()

verbose

defaults to TRUE, use FALSE to silence messages

Value

a dataframe of all releases available within a repository.

Examples

# \donttest{
try({ # wrapped in try block to prevent CRAN errors
 pb_releases("nflverse/nflverse-data")
})
#>          release_name release_id
#> 1        ftn_charting  119740650
#> 2           espn_data  121931789
#> 3      weekly_rosters   73313637
#> 4  players_components  109160206
#> 5             players   69785162
#> 6   pbp_participation   71636554
#> 7           officials   69773595
#> 8                misc   69471165
#> 9                test   66262662
#> 10        draft_picks   66254658
#> 11          contracts   65273753
#> 12        snap_counts   58153010
#> 13            rosters   58152863
#> 14       player_stats   58152881
#> 15       pfr_advstats   58152981
#> 16                pbp   58152862
#> 17      nextgen_stats   58152931
#> 18           injuries   58152949
#> 19       depth_charts   58152948
#> 20            combine   60901499
#>                                                                                                                                                                                            release_body
#> 1                                                                                                                                                                 Charting Data provided by FTNData.com
#> 2                                                                                                                                                                                            ESPN Stats
#> 3                                                                                                                                       Week-level rosters via NFL Shield v2 API, dating back to 2002. 
#> 4                                                                                                                                                            Component files for nflverse players build
#> 5                                                                                                Player information for nflverse. This should be the go-to for position and ID mappings going forward. 
#> 6                                                                                                                                                                 Participation data for plays from NGS
#> 7                                                                                                                             Data about which referees and officials were assigned to specific games. 
#> 8  Various bits of data stored here as onetime jobs. Not automatically updated. Please file an issue or mention it on discord if you'd like to request something be maintained on an automatic schedule
#> 9                                                                                                                                                                                  for internal testing
#> 10                                                        Draft picks dating back to 1980, courtesy of Pro Football Reference. \r\n\r\nIncludes some career-level and ApproxValue/HOF/MVP/PB stats.\r\n
#> 11                                                                                                                                OverTheCap contract data, accessed with `nflreadr::load_contracts()`.
#> 12                                                                                                      Snap counts data, accessed with `nflreadr::load_snap_counts()`.\r\n\r\nLast Updated: 2022-03-06
#> 13                                                                                                                                                Roster data, accessed with `nflreadr::load_rosters()`
#> 14                                                                                                                                     Play by play data, accessed with `nflreadr::load_player_stats()`
#> 15                                                                                                                        PFR Adv Stats data, accessed with `nflreadr::load_pfr_advstats()`\r\n\r\n\r\n
#> 16                                                                                                                                              Play by play data, accessed with `nflreadr::load_pbp()`
#> 17                                                                                                                              NFL Next Gen Stats data, accessed with `nflreadr::load_nextgen_stats()`
#> 18                                                                                                                                             Injuries data, accessed with `nflreadr::load_injuries()`
#> 19                                                                                                                                      Depth chart data, accessed with `nflreadr::load_depth_charts()`
#> 20                                                                                NFL Combine Data courtesy of PFR, accessed with `nflreadr::load_combine()`. \r\n\r\nLast Updated: 2022-03-05 17:06:33
#>              tag_name draft latest           created_at         published_at
#> 1        ftn_charting FALSE  FALSE 2023-09-01T02:16:48Z 2023-09-03T16:14:36Z
#> 2           espn_data FALSE  FALSE 2023-09-01T02:16:48Z 2023-09-20T19:50:55Z
#> 3      weekly_rosters FALSE  FALSE 2022-06-15T16:58:26Z 2022-08-01T07:45:30Z
#> 4  players_components FALSE  FALSE 2022-06-15T16:58:26Z 2023-06-20T01:22:44Z
#> 5             players FALSE  FALSE 2022-06-15T16:58:26Z 2022-06-19T05:36:22Z
#> 6   pbp_participation FALSE  FALSE 2022-06-15T16:58:26Z 2022-07-10T04:32:49Z
#> 7           officials FALSE  FALSE 2022-06-15T16:58:26Z 2022-06-18T17:47:11Z
#> 8                misc FALSE  FALSE 2022-05-29T16:57:43Z 2022-06-14T23:46:04Z
#> 9                test FALSE  FALSE 2022-05-06T19:04:21Z 2022-05-06T19:56:46Z
#> 10        draft_picks FALSE  FALSE 2022-05-06T18:12:18Z 2022-05-06T18:15:12Z
#> 11          contracts FALSE  FALSE 2022-03-21T14:03:51Z 2022-04-25T19:40:32Z
#> 12        snap_counts FALSE  FALSE 2022-01-28T02:09:45Z 2022-01-28T02:15:02Z
#> 13            rosters FALSE  FALSE 2022-01-28T02:09:45Z 2022-01-28T02:12:10Z
#> 14       player_stats FALSE  FALSE 2022-01-28T02:09:45Z 2022-01-28T02:12:36Z
#> 15       pfr_advstats FALSE  FALSE 2022-01-28T02:09:45Z 2022-01-28T02:14:24Z
#> 16                pbp FALSE   TRUE 2022-01-28T02:09:45Z 2022-01-28T02:12:09Z
#> 17      nextgen_stats FALSE  FALSE 2022-01-28T02:09:45Z 2022-01-28T02:13:25Z
#> 18           injuries FALSE  FALSE 2022-01-28T02:09:45Z 2022-01-28T02:13:49Z
#> 19       depth_charts FALSE  FALSE 2022-01-28T02:09:45Z 2022-01-28T02:13:47Z
#> 20            combine FALSE  FALSE 2022-01-28T02:09:45Z 2022-03-03T15:51:29Z
#>                                                                     html_url
#> 1        https://github.com/nflverse/nflverse-data/releases/tag/ftn_charting
#> 2           https://github.com/nflverse/nflverse-data/releases/tag/espn_data
#> 3      https://github.com/nflverse/nflverse-data/releases/tag/weekly_rosters
#> 4  https://github.com/nflverse/nflverse-data/releases/tag/players_components
#> 5             https://github.com/nflverse/nflverse-data/releases/tag/players
#> 6   https://github.com/nflverse/nflverse-data/releases/tag/pbp_participation
#> 7           https://github.com/nflverse/nflverse-data/releases/tag/officials
#> 8                https://github.com/nflverse/nflverse-data/releases/tag/misc
#> 9                https://github.com/nflverse/nflverse-data/releases/tag/test
#> 10        https://github.com/nflverse/nflverse-data/releases/tag/draft_picks
#> 11          https://github.com/nflverse/nflverse-data/releases/tag/contracts
#> 12        https://github.com/nflverse/nflverse-data/releases/tag/snap_counts
#> 13            https://github.com/nflverse/nflverse-data/releases/tag/rosters
#> 14       https://github.com/nflverse/nflverse-data/releases/tag/player_stats
#> 15       https://github.com/nflverse/nflverse-data/releases/tag/pfr_advstats
#> 16                https://github.com/nflverse/nflverse-data/releases/tag/pbp
#> 17      https://github.com/nflverse/nflverse-data/releases/tag/nextgen_stats
#> 18           https://github.com/nflverse/nflverse-data/releases/tag/injuries
#> 19       https://github.com/nflverse/nflverse-data/releases/tag/depth_charts
#> 20            https://github.com/nflverse/nflverse-data/releases/tag/combine
#>                                                                                        upload_url
#> 1  https://uploads.github.com/repos/nflverse/nflverse-data/releases/119740650/assets{?name,label}
#> 2  https://uploads.github.com/repos/nflverse/nflverse-data/releases/121931789/assets{?name,label}
#> 3   https://uploads.github.com/repos/nflverse/nflverse-data/releases/73313637/assets{?name,label}
#> 4  https://uploads.github.com/repos/nflverse/nflverse-data/releases/109160206/assets{?name,label}
#> 5   https://uploads.github.com/repos/nflverse/nflverse-data/releases/69785162/assets{?name,label}
#> 6   https://uploads.github.com/repos/nflverse/nflverse-data/releases/71636554/assets{?name,label}
#> 7   https://uploads.github.com/repos/nflverse/nflverse-data/releases/69773595/assets{?name,label}
#> 8   https://uploads.github.com/repos/nflverse/nflverse-data/releases/69471165/assets{?name,label}
#> 9   https://uploads.github.com/repos/nflverse/nflverse-data/releases/66262662/assets{?name,label}
#> 10  https://uploads.github.com/repos/nflverse/nflverse-data/releases/66254658/assets{?name,label}
#> 11  https://uploads.github.com/repos/nflverse/nflverse-data/releases/65273753/assets{?name,label}
#> 12  https://uploads.github.com/repos/nflverse/nflverse-data/releases/58153010/assets{?name,label}
#> 13  https://uploads.github.com/repos/nflverse/nflverse-data/releases/58152863/assets{?name,label}
#> 14  https://uploads.github.com/repos/nflverse/nflverse-data/releases/58152881/assets{?name,label}
#> 15  https://uploads.github.com/repos/nflverse/nflverse-data/releases/58152981/assets{?name,label}
#> 16  https://uploads.github.com/repos/nflverse/nflverse-data/releases/58152862/assets{?name,label}
#> 17  https://uploads.github.com/repos/nflverse/nflverse-data/releases/58152931/assets{?name,label}
#> 18  https://uploads.github.com/repos/nflverse/nflverse-data/releases/58152949/assets{?name,label}
#> 19  https://uploads.github.com/repos/nflverse/nflverse-data/releases/58152948/assets{?name,label}
#> 20  https://uploads.github.com/repos/nflverse/nflverse-data/releases/60901499/assets{?name,label}
#>    n_assets
#> 1        10
#> 2        10
#> 3        90
#> 4        10
#> 5         6
#> 6        34
#> 7         6
#> 8        17
#> 9         2
#> 10        6
#> 11        7
#> 12       51
#> 13      418
#> 14      392
#> 15      138
#> 16      150
#> 17       86
#> 18       62
#> 19       94
#> 20        6
# }